♻️ Export individual functions, precise typing
