+++ packages/hub/package.json
@@ -49,7 +49,7 @@
 	"license": "MIT",
 	"devDependencies": {
 		"@types/node": "^18.13.0",
-		"typescript": "^4.9.5"
+		"typescript": "^5.0.4"
 	},
 	"dependencies": {
 		"hash-wasm": "^4.9.0",
