+++ packages/inference/package.json
@@ -51,7 +51,7 @@
 	},
 	"devDependencies": {
 		"@types/node": "18.13.0",
-		"typescript": "4.9.5",
+		"typescript": "^5.0.4",
 		"vite": "^4.1.4",
 		"vitest": "^0.29.8"
 	},
