+++ packages/inference/src/index.ts
@@ -1,1 +1,3 @@
-export * from "./HfInference";
+export { HfInference, HfInferenceEndpoint } from "./HfInference";
+export type * from "./types";
+export * from "./lib";
