+++ packages/inference/src/lib/imageClassification.ts
@@ -1,0 +1,19 @@
+import type { ImageClassificationArgs, ImageClassificationReturn, Options } from "../types";
+import { request } from "./request";
+
+/**
+ * This task reads some image input and outputs the likelihood of classes.
+ * Recommended model: google/vit-base-patch16-224
+ */
+export async function imageClassification(
+	args: ImageClassificationArgs,
+	options?: Options
+): Promise<ImageClassificationReturn> {
+	const res = await request<ImageClassificationReturn>(args, options);
+	const isValidOutput =
+		Array.isArray(res) && res.every((x) => typeof x.label === "string" && typeof x.score === "number");
+	if (!isValidOutput) {
+		throw new TypeError("Invalid inference output: output must be of type Array<label: string, score: number>");
+	}
+	return res;
+}
