+++ packages/inference/src/lib/index.ts
@@ -1,0 +1,22 @@
+export { audioClassification } from "./audioClassification";
+export { automaticSpeechRecognition } from "./automaticSpeechRecognition";
+export { conversational } from "./conversational";
+export { featureExtraction } from "./featureExtraction";
+export { fillMask } from "./fillMask";
+export { imageClassification } from "./imageClassification";
+export { imageSegmentation } from "./imageSegmentation";
+export { imageToText } from "./imageToText";
+export { objectDetection } from "./objectDetection";
+export { questionAnswering } from "./questionAnswering";
+export { request } from "./request";
+export { sentenceSimilarity } from "./sentenceSimilarity";
+export { streamingRequest } from "./streamingRequest";
+export { summarization } from "./summarization";
+export { tableQuestionAnswering } from "./tableQuestionAnswering";
+export { textClassification } from "./textClassification";
+export { textGeneration } from "./textGeneration";
+export { textGenerationStream } from "./textGenerationStream";
+export { textToImage } from "./textToImage";
+export { tokenClassification } from "./tokenClassification";
+export { translation } from "./translation";
+export { zeroShotClassification } from "./zeroShotClassification";
